Pass the cluster to the partitioner
We probably need to cache the cluster instance because wrapping it on every call is going to be expensive...

I've added PartitionInfo#eql? to be able to write tests that show that what you get from the cluster is the same as what you get from other metadata calls (in one sense this tests Kafka more than the driver code, but it makes sure that all the wrapper objects behave as they should).