Add new unit tests for Consumer#seek_to_{end,beginning} + support for no arg
Turns out I had missed that you're allowed to pass an empty collection to #seek_to_{end,beginning} to make it seek for all assigned partitions. I decided that this would be equivalent to making the list of partitions an optional argument.

Unfortunately MockConsumer doesn't implement this detail of the Consumer API.